requirements.txt
mmcv-full==1.4.0
numpy 
opencv-python==4.5.5.64
Pillow==9.0.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111